% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wireShiny.R
\name{run_wireShiny}
\alias{run_wireShiny}
\title{Run the Shiny app based on functionalities of \code{wire}}
\usage{
run_wireShiny()
}
\description{
Run the Shiny app based on functionalities of \code{wire}
}
\examples{
if (interactive()) {
  run_wireShiny()
}

}
