% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x3p_subsamples}
\alias{x3p_subsamples}
\title{Subsampled dataset of wire cuts}
\format{
A list of \code{x3p} objects, each representing a surface scan. The list includes:
\itemize{
\item T1AW-LI-R1: An \code{x3p} object representing the first surface scan.
\item T1AW-LI-R2: An \code{x3p} object representing the second surface scan.
}
}
\usage{
x3p_subsamples
}
\description{
This dataset is a subsampled version of the original \code{x3p} wire cuts data. The subsampling is performed at a factor of 10.
}
\keyword{datasets}
